19-04-2022 16:55

+++ biblio/src/main/java/com/example/biblio/models/Auteur.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class Auteur {
     @Id
     @Column(name = "idAuteur", nullable = false)
-    private int idAuteur;
+    private Long idAuteur;
     @Column(name = "nomAuteur", nullable = false, length = 200)
     private String nomAuteur;
     @Column(name = "nationaliteAuteur", nullable = false, length = 50)
